Split `analysis.rs` into multiple modules
